Move normaliseDataset in its own module
This'll allow the polyfill it requires to be tree-shaken properly when bundling components
that require common.mjs
