+++ src/govuk/components/accordion/accordion.mjs
@@ -1,8 +1,9 @@
-import { nodeListForEach, mergeConfigs, extractConfigByNamespace, normaliseDataset } from '../../common.mjs'
+import { nodeListForEach, mergeConfigs, extractConfigByNamespace } from '../../common.mjs'
 import { I18n } from '../../i18n.mjs'
 import '../../vendor/polyfills/Function/prototype/bind.mjs'
 import '../../vendor/polyfills/Element/prototype/classList.mjs'
 import '../../vendor/polyfills/String/prototype/trim.mjs'
+import { normaliseDataset } from '../../common/normalise-dataset.mjs'
 
 /**
  * @type {AccordionTranslations}
