+++ src/govuk/components/error-summary/error-summary.mjs
@@ -2,7 +2,8 @@ import '../../vendor/polyfills/Function/prototype/bind.mjs'
 import '../../vendor/polyfills/Event.mjs' // addEventListener
 import '../../vendor/polyfills/Element/prototype/closest.mjs'
 
-import { mergeConfigs, normaliseDataset } from '../../common.mjs'
+import { mergeConfigs } from '../../common.mjs'
+import { normaliseDataset } from '../../common/normalise-dataset.mjs'
 
 /**
  * JavaScript enhancements for the ErrorSummary
